Split examples into smaller classes
